Restructure the test directory to mirror the src directory
