+++ test/helpers/map.spec.js
@@ -7,9 +7,9 @@ import {
   getTileTraversabilityInCells,
   getMapTraversabilityInCells,
   __RewireAPI__ as MapRewireAPI,
-} from '../src/helpers/map';
-import { setupTiles, teardownTiles } from './tiles.spec';
-import { getTileId } from '../src/tiles';
+} from '../../src/helpers/map';
+import { setupTiles, teardownTiles } from '../tiles.spec';
+import { getTileId } from '../../src/tiles';
 
 
 test('init2dArray: returns 2d array that is correct size, and with correct value filled in', t => {
