+++ test/utils/asserts.spec.js
@@ -1,5 +1,5 @@
 import test from 'tape';
-import { assert, assertGridInBounds } from '../src/utils/asserts';
+import { assert, assertGridInBounds } from '../../src/utils/asserts';
 
 
 test('assert: does not throw errors with condition=true', t => {
